remove v2 dup'd logic
