+++ pkg/integrations/install/install.go
@@ -37,7 +37,6 @@ import (
 	_ "github.com/grafana/agent/pkg/integrations/v2/app_agent_receiver" // register app_agent_receiver
 	_ "github.com/grafana/agent/pkg/integrations/v2/blackbox_exporter"  // register blackbox_exporter
 	_ "github.com/grafana/agent/pkg/integrations/v2/eventhandler"       // register eventhandler
-	_ "github.com/grafana/agent/pkg/integrations/v2/mssql"              // register mssql
 	_ "github.com/grafana/agent/pkg/integrations/v2/snmp_exporter"      // register snmp_exporter
 	_ "github.com/grafana/agent/pkg/integrations/v2/snowflake"          // register snowflake
 	_ "github.com/grafana/agent/pkg/integrations/v2/vmware_exporter"    // register vmware_exporter
